+++ packages/client/src/App/views/Bots/Viewer.tsx
@@ -1,14 +1,13 @@
 import { SimpleGrid, Stack, Heading, GridItem } from '@chakra-ui/layout'
 import { Redirect, useLocation } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
-import { Button } from '@chakra-ui/button'
 import { Bot } from 'atom'
 
 import BotControl from 'components/bots/controls'
 import UIBreadcrumb from 'components/ui/breadcrumb'
-import { UICard } from 'components/ui/card'
 import ListModules from 'components/bots/list-modules'
 import ListCommands from 'components/bots/list-commands'
+import ListRoles from 'components/bots/list-roles'
 
 const BotViewer: React.FC = () => {
 	const { t } = useTranslation()
@@ -38,8 +37,13 @@ const BotViewer: React.FC = () => {
 				</GridItem>
 
 				<GridItem colSpan={{ base: 10, md: 5, '2xl': 4 }}>
-					<ListCommands />
+					<Stack>
+						<ListCommands />
+						<ListRoles />
+					</Stack>
 				</GridItem>
+
+				<GridItem colSpan={{ base: 10, md: 5, '2xl': 4 }}></GridItem>
 			</SimpleGrid>
 		</Stack>
 	)