+++ packages/client/src/components/bots/list-modules.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { SimpleGrid, GridItem } from '@chakra-ui/layout'
+import { SimpleGrid, GridItem, Center, Text } from '@chakra-ui/layout'
 import { Redirect, Link, useLocation } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import { Button, ButtonGroup } from '@chakra-ui/button'
@@ -8,9 +8,11 @@ import { Bot, InstalledModule, modulesAtom } from 'atom'
 
 import { UICard } from 'components/ui/card'
 import realsync from 'providers/realsync'
+import { Spinner } from '@chakra-ui/spinner'
 
 const ListModules: React.FC = () => {
 	const { t } = useTranslation()
+	const [loading, setLoading] = useState(true)
 	const { state: bot, pathname } = useLocation<Bot>()
 	const [installedModules, setInstallModules] = useState<InstalledModule[]>([])
 	const [modules] = useRecoilState(modulesAtom)
@@ -18,6 +20,7 @@ const ListModules: React.FC = () => {
 	const Fetch = async () => {
 		const data = await realsync.service('module/active', bot.id)
 		setInstallModules(data as InstalledModule[])
+		setLoading(false)
 	}
 
 	useEffect(() => {
@@ -27,6 +30,22 @@ const ListModules: React.FC = () => {
 
 	if (!bot) {
 		return <Redirect to='/home' />
+	}
+
+	if (loading) {
+		return (
+			<Center>
+				<Spinner />
+			</Center>
+		)
+	}
+
+	if (!loading && installedModules.length == 0) {
+		return (
+			<Center>
+				<Text>No modules found</Text>
+			</Center>
+		)
 	}
 
 	return (