+++ packages/client/src/components/bots/list-roles.tsx
@@ -1,0 +1,14 @@
+import { Box, Text } from '@chakra-ui/layout'
+import { UICard } from 'components/ui/card'
+
+const ListRoles: React.FC = () => {
+	return (
+		<Box>
+			<UICard title='Roles'>
+				<Text>Coming soon</Text>
+			</UICard>
+		</Box>
+	)
+}
+
+export default ListRoles